LayoutTests/inspector/debugger/script-formatter-expected.txt

CONSOLE MESSAGE: line 24: Test message.
Tests the script formatting functionality.

Debugger was enabled.

Running: testScriptFormatterWorker

Running: testSourceMapping

Running: testFormatInlinedScripts
<html><body><script>
function f() {
}
</script><script>
function g() {
}
</script></body></html>

Running: testBreakpointsInOriginalAndFormattedSource
Script execution paused.
Call stack:
    0) f1 (script-formatter.html:12)
    1)  (script-formatter.html:1)
Script execution resumed.
Page reloaded.
Navigated to test-navigation.html
Page reloaded.
Navigated back to test page.
Script execution paused.
Call stack:
    0) f2 (script-formatter.html:17)
    1)  (script-formatter.html:1)
Script execution resumed.

Running: testConsoleMessagesForFormattedScripts
script-formatter.html:24
script-formatter.html:32
Debugger was disabled.
